+++ modules/template/elements/UsersElement.php
@@ -8,10 +8,10 @@
 
 namespace humhub\modules\custom_pages\modules\template\elements;
 
-use humhub\modules\ui\form\widgets\ActiveForm;
 use humhub\modules\user\models\Group;
 use humhub\modules\user\models\User;
 use humhub\modules\user\widgets\UserPickerField;
+use humhub\widgets\form\ActiveForm;
 use Yii;
 use yii\db\ActiveQuery;