+++ modules/template/views/admin/edit.php
@@ -6,11 +6,11 @@
  * @license https://www.humhub.com/licences
  */
 
-use humhub\libs\Html;
-use humhub\modules\ui\form\widgets\ActiveForm;
-use humhub\modules\ui\view\components\View;
-use humhub\widgets\Button;
+use humhub\components\View;
+use humhub\helpers\Html;
 use humhub\modules\custom_pages\modules\template\models\Template;
+use humhub\widgets\bootstrap\Button;
+use humhub\widgets\form\ActiveForm;
 
 /* @var Template $model */
 /* @var View $this */
@@ -28,7 +28,7 @@ use humhub\modules\custom_pages\modules\template\models\Template;
         <?= $form->field($model, 'description')->textarea(['id' => 'template-form-description', 'rows' => 3]) ?>
         <?= $form->field($model, 'type')->dropDownList($model::getTypeOptions()) ?>
 
-        <div id="template-allow-for-spaces"<?= $model->isLayout() ? '' : ' style="display:none"'?>>
+        <div id="template-allow-for-spaces"<?= $model->isLayout() ? '' : ' class="d-none"' ?>">
             <?= $form->field($model, 'allow_for_spaces')->checkbox() ?>
         </div>