+++ modules/template/widgets/views/templateElementAdminRow.php
@@ -7,8 +7,8 @@
  */
 
 use humhub\modules\custom_pages\modules\template\models\TemplateElement;
-use humhub\widgets\Button;
-use humhub\widgets\Label;
+use humhub\widgets\bootstrap\Badge;
+use humhub\widgets\bootstrap\Button;
 use yii\helpers\Html;
 
 /* @var $model TemplateElement */
@@ -21,26 +21,26 @@ use yii\helpers\Html;
         {{ <?= Html::encode($model->name) ?> }}
     </td>
     <td>
-        <?= Label::success($model->getLabel()) ?>
+        <?= Badge::success($model->getLabel()) ?>
         <?php if (!$model->hasDefaultContent()) : ?>
-            <?= Label::warning(Yii::t('CustomPagesModule.base', 'Empty')) ?>
-        <?php else : ?>
-            <?= Label::success(Yii::t('CustomPagesModule.base', 'Default')) ?>
+            <?= Badge::warning(Yii::t('CustomPagesModule.base', 'Empty')) ?>
+        <?php else: ?>
+            <?= Badge::success(Yii::t('CustomPagesModule.base', 'Default')) ?>
         <?php endif; ?>
     </td>
     <td>
     <?php if ($model->template->canEdit()) : ?>
-        <?= Button::danger()->icon('times')->xs()
+        <?= Button::danger()->icon('times')->sm()
             ->action('deleteElementSubmit', ['/custom_pages/template/admin/delete-element', 'id' => $model->id])
             ->confirm(
                 Yii::t('CustomPagesModule.template', '<strong>Confirm</strong> element deletion'),
                 Yii::t('CustomPagesModule.template', 'Do you really want to delete this element? <br />The deletion will affect all pages using this template.'),
                 Yii::t('CustomPagesModule.base', 'Delete'),
             ) ?>
-        <?= Button::primary()->icon('pencil')->xs()
+        <?= Button::primary()->icon('pencil')->sm()
             ->action('ui.modal.load', ['/custom_pages/template/admin/edit-element', 'id' => $model->id]) ?>
     <?php else : ?>
-        <?= Button::info()->icon('eye')->xs()
+        <?= Button::accent()->icon('eye')->sm()
             ->action('ui.modal.load', ['/custom_pages/template/admin/edit-element', 'id' => $model->id]) ?>
     <?php endif; ?>
     </td>