+++ modules/template/widgets/views/templateStructure.php
@@ -6,19 +6,17 @@
  * @license https://www.humhub.com/licences
  */
 
-use humhub\libs\Html;
+use humhub\components\View;
+use humhub\helpers\Html;
 use humhub\modules\content\helpers\ContentContainerHelper;
-use humhub\modules\custom_pages\assets\Assets;
 use humhub\modules\custom_pages\helpers\Url;
-use humhub\modules\custom_pages\modules\template\assets\InlineEditorAsset;
 use humhub\modules\custom_pages\modules\template\elements\ContainerElement;
 use humhub\modules\custom_pages\modules\template\elements\ContainerItem;
 use humhub\modules\custom_pages\modules\template\helpers\PagePermissionHelper;
 use humhub\modules\custom_pages\modules\template\models\TemplateInstance;
 use humhub\modules\custom_pages\modules\template\widgets\TemplateStructure;
 use humhub\modules\ui\icon\widgets\Icon;
-use humhub\modules\ui\view\components\View;
-use humhub\widgets\Link;
+use humhub\widgets\bootstrap\Link;
 
 /* @var TemplateInstance $templateInstance */
 /* @var ContainerElement[] $containers */
@@ -29,25 +27,21 @@ use humhub\widgets\Link;
 
 /* @var TemplateStructure $widget */
 $widget = $this->context;
-
-Assets::register($this);
-InlineEditorAsset::register($this);
 ?>
 <?php if ($templateInstance->isPage()) : ?>
 <?= Html::beginTag('div', $options) ?>
     <div class="cp-structure-header cp-structure-row">
         <div class="cp-structure-actions dropdown">
-            <?= Icon::get('ellipsis-h', ['htmlOptions' => ['data-toggle' => 'dropdown']])
-                ->class('dropdown-toggle') ?>
+            <?= Icon::get('ellipsis-h', ['htmlOptions' => ['data-bs-toggle' => 'dropdown']]) ?>
             <ul class="dropdown-menu">
                 <li><?= Link::to(
                     Yii::t('CustomPagesModule.view', 'Page configuration'),
                     Url::toEditPage($templateInstance->page_id, ContentContainerHelper::getCurrent()),
-                )->blank() ?></li>
+                )->blank()->cssClass('dropdown-item') ?></li>
                 <li><?= Link::to(
                     Yii::t('CustomPagesModule.view', 'Exit Edit Mode'),
                     Url::toViewPage($templateInstance->page_id, ContentContainerHelper::getCurrent()),
-                )->pjax(false) ?></li>
+                )->cssClass('dropdown-item')->pjax(false) ?></li>
             </ul>
         </div>
         <div class="cp-structure-text"><?= Yii::t('CustomPagesModule.template', 'Structure View') ?></div>
@@ -61,22 +55,26 @@ InlineEditorAsset::register($this);
             <?= Icon::get('circle') ?>
             <div class="cp-structure-text"><?= $templateInstance->template->name ?></div>
             <div class="cp-structure-actions dropdown">
-                <?= Icon::get('ellipsis-h', ['htmlOptions' => ['data-toggle' => 'dropdown']])
-                    ->class('dropdown-toggle cp-structure-action') ?>
-                <ul class="dropdown-menu dropdown-menu-right">
+                <?= Icon::get('ellipsis-h', ['htmlOptions' => ['data-bs-toggle' => 'dropdown']])
+                    ->class('cp-structure-action') ?>
+                <ul class="dropdown-menu dropdown-menu-end">
                 <?php if ($templateInstance->isContainer()) : ?>
                     <li><?= Link::to(Yii::t('CustomPagesModule.template', 'Edit'))
                         ->icon('pencil')
-                        ->action('editElements') ?></li>
+                        ->action('editElements')
+                        ->cssClass('dropdown-item') ?></li>
                     <li><?= Link::to(Yii::t('CustomPagesModule.template', 'Move Up'))
                         ->icon('chevron-up')
-                        ->action('moveUpContainerItem') ?></li>
+                        ->action('moveUpContainerItem')
+                            ->cssClass('dropdown-item') ?></li>
                     <li><?= Link::to(Yii::t('CustomPagesModule.template', 'Move Down'))
                         ->icon('chevron-down')
-                        ->action('moveDownContainerItem') ?></li>
+                        ->action('moveDownContainerItem')
+                            ->cssClass('dropdown-item') ?></li>
                     <li><?= Link::to(Yii::t('CustomPagesModule.template', 'Export'))
                             ->icon('upload')
-                            ->action('exportTemplateInstance') ?></li>
+                            ->action('exportTemplateInstance')
+                            ->cssClass('dropdown-item') ?></li>
                     <li><?= Link::to(Yii::t('CustomPagesModule.template', 'Delete'))
                         ->icon('trash')
                         ->action('deleteContainerItem')
@@ -84,24 +82,29 @@ InlineEditorAsset::register($this);
                             Yii::t('CustomPagesModule.template', '<strong>Confirm</strong> container item deletion'),
                             Yii::t('CustomPagesModule.template', 'Are you sure you want to delete this container item?'),
                             Yii::t('CustomPagesModule.base', 'Delete'),
-                        ) ?></li>
+                        )
+                        ->cssClass('dropdown-item') ?></li>
                 <?php else : ?>
                     <li><?= Link::to(Yii::t('CustomPagesModule.template', 'Edit'))
                         ->icon('pencil')
-                        ->action('editElements') ?></li>
+                        ->action('editElements')
+                        ->cssClass('dropdown-item') ?></li>
                     <li><?= Link::to(Yii::t('CustomPagesModule.template', 'Import'))
                             ->icon('download')
-                            ->action('importTemplateInstance') ?></li>
+                            ->action('importTemplateInstance')
+                            ->cssClass('dropdown-item') ?></li>
                     <li><?= Link::to(Yii::t('CustomPagesModule.template', 'Export'))
                         ->icon('upload')
-                        ->action('exportTemplateInstance') ?></li>
+                        ->action('exportTemplateInstance')
+                        ->cssClass('dropdown-item') ?></li>
                 <?php endif; ?>
 
                 <?php if (PagePermissionHelper::canTemplate()) : ?>
                     <li><?= Link::to(Yii::t('CustomPagesModule.template', 'Edit template'))
                             ->icon('pencil-square-o')
                             ->link(['/custom_pages/template/admin/edit-source', 'id' => $templateInstance->template_id])
-                            ->blank() ?></li>
+                            ->blank()
+                            ->cssClass('dropdown-item') ?></li>
                 <?php endif; ?>
                 </ul>
             </div>
@@ -116,20 +119,21 @@ InlineEditorAsset::register($this);
                     <div class="cp-structure-text"><?= $container->element->title === null || $container->element->title === '' ? $container->element->name : $container->element->title ?></div>
 
                     <div class="cp-structure-actions dropdown">
-                        <?= Icon::get('ellipsis-h', ['htmlOptions' => ['data-toggle' => 'dropdown']])
-                            ->class('dropdown-toggle cp-structure-action') ?>
-                        <ul class="dropdown-menu dropdown-menu-right">
+                        <?= Icon::get('ellipsis-h', ['htmlOptions' => ['data-bs-toggle' => 'dropdown']])
+                            ->class('cp-structure-action') ?>
+                        <ul class="dropdown-menu dropdown-menu-end">
                             <li><?= Link::to(Yii::t('CustomPagesModule.template', 'Import'))
                                     ->icon('download')
-                                    ->action('importTemplateInstance') ?></li>
+                                    ->action('importTemplateInstance')
+                                    ->cssClass('dropdown-item') ?></li>
                             <li><?= Link::to(Yii::t('CustomPagesModule.template', 'Export'))
                                     ->icon('upload')
-                                    ->action('exportTemplateInstance') ?></li>
+                                    ->action('exportTemplateInstance')
+                                    ->cssClass('dropdown-item') ?></li>
                         </ul>
                     </div>
                     <?= Icon::get('plus', ['htmlOptions' => ['data-action-click' => 'addContainerItem']])
-                        ->class('cp-structure-action')
-                        ->style($container->canAddItem() ? '' : 'display:none') ?>
+                        ->class('cp-structure-action' . ($container->canAddItem() ? '' : ' d-none')) ?>
                 </div>
 
                 <?php if ($container->hasItems()) : ?>