+++ tests/codeception/acceptance/CreateGlobalPageCest.php
@@ -26,10 +26,12 @@ class CreateGlobalPageCest
 
         $I->fillField('CustomPage[title]', 'Test title');
         $I->fillField('#custompage-page_content .humhub-ui-richtext', 'Test Content');
-        $I->jsShow('.form-collapsible-fields.closed fieldset');
+        $I->jsClick('.form-collapsible-fields.closed label');
         $I->fillField('CustomPage[sort_order]', '400');
         $I->selectOption('CustomPage[icon]', ['value' => 'fa-adn']);
 
+        $I->scrollToBottom();
+        $I->wait(1);
         $I->click('Create');
         $I->waitForElementVisible('#topbar-second .fa-adn');
         $I->expectTo('see my new page in the top navigation');
@@ -61,9 +63,12 @@ class CreateGlobalPageCest
 
         $I->fillField('CustomPage[title]', 'Test link');
         $I->fillField('CustomPage[page_content]', '/dashboard/dashboard');
-        $I->jsShow('.form-collapsible-fields.closed fieldset');
+        $I->jsClick('.form-collapsible-fields.closed label');
         $I->fillField('CustomPage[sort_order]', '400');
         $I->selectOption('CustomPage[icon]', ['value' => 'fa-adn']);
+
+        $I->scrollToBottom();
+        $I->wait(1);
         $I->click('Create');
         $I->wait(1);
         $I->amOnRoute(['/user/account/edit']);
@@ -89,6 +94,8 @@ class CreateGlobalPageCest
         $I->see('People Buttons');
         $I->seeElement('.target-page-list.PeopleButtonsWidget');
 
+        $I->scrollTo('.target-page-list.PeopleButtonsWidget');
+        $I->wait(1);
         $I->click('.btn-success', '.target-page-list.PeopleButtonsWidget');
 
         $I->waitForText('Add new page');
@@ -98,10 +105,12 @@ class CreateGlobalPageCest
 
         $I->fillField('CustomPage[title]', 'Custom people page');
         $I->fillField('#custompage-page_content .humhub-ui-richtext', 'Custom people page content');
-        $I->jsShow('.form-collapsible-fields.closed fieldset');
+        $I->jsClick('.form-collapsible-fields.closed label');
         $I->fillField('CustomPage[sort_order]', '200');
         $I->selectOption('CustomPage[icon]', ['value' => 'fa-anchor']);
 
+        $I->scrollToBottom();
+        $I->wait(1);
         $I->click('Create');
         $I->waitForText('People Buttons');
         $I->see('Custom people page');
@@ -109,7 +118,7 @@ class CreateGlobalPageCest
         $I->waitForElementVisible('.container-people .panel-heading .fa-anchor');
         $I->expectTo('see my new page in the people heading buttons');
 
-        $I->click('.container-people .panel-heading .fa-anchor');
+        $I->jsClick('.container-people .panel-heading .fa-anchor');
         $I->expectTo('see no my new page content');
 
         $I->waitForText('Custom people page content');